Split CaC project compiler files
